backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

# AI/ML Models
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
ultralytics==8.0.196  # YOLOv8
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2

# Computer Vision & Embeddings
clip-by-openai==1.1
faiss-cpu==1.7.4  # or faiss-gpu for GPU support
sentence-transformers==2.2.2

# Multimodal AI
llava==1.0.0
llama-cpp-python==0.2.11

# Voice Processing
openai-whisper==20231117
groq==0.4.1

# External APIs
googlemaps==4.10.0
boto3==1.34.0  # AWS S3
supabase==2.0.0

# Vector Database
faiss==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
